定时任务 处理超时订单

+++ toucan-shopping-cloud/services/toucan-shopping-stock/src/main/java/com/toucan/shopping/cloud/stock/controller/ProductSkuStockLockController.java
@@ -290,11 +290,11 @@ public class ProductSkuStockLockController {
             }
             productSkuStockLockVO.setType(null);
             List<ProductSkuStockLockVO> skuStockLockVO = productSkuStockLockService.queryListByVO(productSkuStockLockVO);
-
-            int ret = productSkuStockLockService.deletes(skuStockLockVO.stream().map(ProductSkuStockLockVO::getId).collect(Collectors.toList()));
-            if(ret<=0||ret!=skuStockLockVO.size())
-            {
-                throw new IllegalArgumentException("删除锁定库存出现异常");
+            if(!CollectionUtils.isEmpty(skuStockLockVO)) {
+                int ret = productSkuStockLockService.deletes(skuStockLockVO.stream().map(ProductSkuStockLockVO::getId).collect(Collectors.toList()));
+                if (ret <= 0 || ret != skuStockLockVO.size()) {
+                    throw new IllegalArgumentException("删除锁定库存出现异常");
+                }
             }
         }catch(Exception e)
         {
